# back/central-reserve/shared/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  response.BusinessRoleAssignmentResponse:
    properties:
      business_id:
        type: integer
      business_name:
        type: string
      role_id:
        type: integer
      role_name:
        type: string
    type: object
...
  response.UserCreatedResponse:
    properties:
      email:
        type: string
      message:
        type: string
      password:
        type: string
      success:
        type: boolean
    type: object
  response.UserErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  response.UserListResponse:
    properties:
      count:
        type: integer
      data:
        items:
          $ref: '#/definitions/response.UserResponse'
        type: array
      success:
        type: boolean
    type: object
...
  response.UserResponse:
    properties:
      avatar_url:
        type: string
      business_role_assignments:
        items:
          $ref: '#/definitions/response.BusinessRoleAssignmentResponse'
        type: array
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      is_super_user:
        type: boolean
      last_login_at:
        type: string
      name:
        type: string
      phone:
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:3050
info:
  contact:
    email: backend@example.com
    name: Equipo de Backend
  description: Servicio REST para la gestión de reservas multi-restaurante.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://ejemplo.com/terminos
  title: Restaurant Reservation API
  version: "1.0"
paths:
...
  /businesses/{id}/configured-resources:
    get:
      consumes:
      - application/json
      description: Obtiene la configuración de recursos de un business específico
        por su ID, incluyendo todos los recursos asociados y su estado (activo/inactivo)
      parameters:
      - description: ID del business
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Configuración de recursos del business obtenida exitosamente
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Parámetros inválidos
          schema:
            additionalProperties: true
            type: object
        "401":
          description: No autorizado
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Sin permisos
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Business no encontrado
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Error interno del servidor
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Obtener configuración de recursos de un business por ID
      tags:
      - Business Resources
...
  /users:
    get:
      consumes:
      - application/json
      description: Obtiene la lista filtrada y paginada de usuarios del sistema con
        sus roles y businesses
      parameters:
      - default: 1
        description: Número de página
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 10
        description: Tamaño de página
        in: query
        maximum: 100
        minimum: 1
        name: page_size
        type: integer
      - description: Filtrar por nombre (búsqueda parcial)
        in: query
        name: name
        type: string
      - description: Filtrar por email (búsqueda parcial)
        in: query
        name: email
        type: string
      - description: Filtrar por teléfono (búsqueda parcial)
        in: query
        name: phone
        type: string
      - description: 'Filtrar por IDs de usuarios separados por comas (ej: 1,2,3)'
        in: query
        name: user_ids
        type: string
      - description: Filtrar por estado activo
        in: query
        name: is_active
        type: boolean
      - description: Filtrar por ID de rol
        in: query
        name: role_id
        type: integer
      - description: Filtrar por ID de business
        in: query
        name: business_id
        type: integer
      - description: Filtrar por fecha de creación (YYYY-MM-DD o YYYY-MM-DD,YYYY-MM-DD
          para rango)
        in: query
        name: created_at
        type: string
      - default: created_at
        description: Campo para ordenar
        enum:
        - id
        - name
        - email
        - phone
        - is_active
        - created_at
        - updated_at
        in: query
        name: sort_by
        type: string
      - default: desc
        description: Orden de clasificación
        enum:
        - asc
        - desc
        in: query
        name: sort_order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Usuarios obtenidos exitosamente
          schema:
            $ref: '#/definitions/response.UserListResponse'
        "400":
          description: Parámetros de filtro inválidos
          schema:
            $ref: '#/definitions/response.UserErrorResponse'
        "401":
          description: Token de acceso requerido
          schema:
            $ref: '#/definitions/response.UserErrorResponse'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/response.UserErrorResponse'
      security:
      - BearerAuth: []
      summary: Obtener usuarios filtrados y paginados
      tags:
      - Users
    post:
      consumes:
      - multipart/form-data
      description: |-
        Crea un nuevo usuario.
        ENVÍO: SOLO multipart/form-data (no JSON body)
      parameters:
      - description: Nombre (2-100)
        in: formData
        name: name
        required: true
        type: string
      - description: Email válido
        in: formData
        name: email
        required: true
        type: string
      - description: Teléfono (exactamente 10 dígitos)
        in: formData
        name: phone
        type: string
      - description: ¿Activo? (true/false)
        in: formData
        name: is_active
        type: boolean
      - description: Imagen de avatar (sube a S3)
        in: formData
        name: avatarFile
        type: file
      - description: 'IDs de negocios separados por comas. Ej: ''16,21'''
        in: formData
        name: business_ids
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Usuario creado exitosamente (incluye contraseña generada)
          schema:
            $ref: '#/definitions/response.UserCreatedResponse'
        "400":
          description: Datos inválidos (mensajes claros de validación)
          schema:
            $ref: '#/definitions/response.UserErrorResponse'
        "401":
          description: Token de acceso requerido
          schema:
            $ref: '#/definitions/response.UserErrorResponse'
        "409":
          description: Email ya existe
          schema:
            $ref: '#/definitions/response.UserErrorResponse'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/response.UserErrorResponse'
      security:
      - BearerAuth: []
      summary: Crear usuario
      tags:
      - Users
...
securityDefinitions:
  BearerAuth:
    description: Token de business JWT con el prefijo **Bearer** (para todos los endpoints)
    in: header
    name: Authorization
    type: apiKey
  BusinessTokenAuth:
    description: Token principal JWT con el prefijo **Bearer** (solo para /auth/business-token)
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"